requirements.txt
langgraph>=0.0.55
langchain-groq>=0.1.5
langchain-core>=0.2.0
langchain>=0.2.0

# sqlite3
pandas>=2.0.0
sqlalchemy>=2.0.0


python-dotenv>=1.0.0
streamlit>=1.28.0 
rich>=13.0.0      

pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0